+++ src/Connection/Protocols/LegacyProtocol.php
@@ -181,7 +181,15 @@ class LegacyProtocol extends Protocol {
      * @throws RuntimeException
      */
     public function selectFolder(string $folder = 'INBOX') {
-        \imap_reopen($this->stream, $this->getAddress().$folder, IMAP::OP_READONLY, 3);
+        $flags = IMAP::OP_READONLY;
+        if (in_array($this->protocol, ["pop3", "nntp"])) {
+            $flags = IMAP::NIL;
+        }
+        if ($this->stream === false) {
+            throw new RuntimeException("failed to reopen stream.");
+        }
+
+        \imap_reopen($this->stream, $this->getAddress().$folder, $flags, 3);
         $this->uid_cache = null;
         return $this->examineFolder($folder);
     }
@@ -412,7 +420,7 @@ class LegacyProtocol extends Protocol {
      */
     public function copyManyMessages(array $messages, string $folder, $uid = IMAP::ST_UID) {
         foreach($messages as $msg) {
-            if ($this->copyMessage($folder, $msg, null, $uid) == false) {
+            if (!$this->copyMessage($folder, $msg, null, $uid)) {
                 return false;
             }
         }
@@ -421,7 +429,7 @@ class LegacyProtocol extends Protocol {
     }
 
     /**
-     * Move a message set from current folder to an other folder
+     * Move a message set from current folder to another folder
      * @param string $folder destination folder
      * @param $from
      * @param int|null $to if null only one message ($from) is fetched, else it's the
@@ -444,7 +452,7 @@ class LegacyProtocol extends Protocol {
      */
     public function moveManyMessages(array $messages, string $folder, $uid = IMAP::ST_UID) {
         foreach($messages as $msg) {
-            if ($this->moveMessage($folder, $msg, null, $uid) == false) {
+            if (!$this->moveMessage($folder, $msg, null, $uid)) {
                 return false;
             }
         }