+++ src/Folder.php
@@ -278,7 +278,7 @@ class Folder {
      */
     public function appendMessage(string $message, array $options = null, $internal_date = null): bool {
         /**
-         * Check if $internal_date is parsed. If it is null it should not be set. Otherwise the message can't be stored.
+         * Check if $internal_date is parsed. If it is null it should not be set. Otherwise, the message can't be stored.
          * If this parameter is set, it will set the INTERNALDATE on the appended message. The parameter should be a
          * date string that conforms to the rfc2060 specifications for a date_time value or be a Carbon object.
          */
@@ -351,8 +351,8 @@ class Folder {
     /**
      * Idle the current connection
      * @param callable $callback
-     * @param integer $timeout max 1740 seconds - recommended by rfc2177 §3
-     * @param boolean $auto_reconnect try to reconnect on connection close
+     * @param integer $timeout max 1740 seconds - recommended by rfc2177 §3. Should not be lower than the servers "* OK Still here" message interval
+     * @param boolean $auto_reconnect try to reconnect on connection close (@deprecated is no longer required)
      *
      * @throws ConnectionFailedException
      * @throws Exceptions\InvalidMessageDateException
@@ -364,25 +364,25 @@ class Folder {
      * @throws Exceptions\MessageNotFoundException
      * @throws Exceptions\NotSupportedCapabilityException
      */
-    public function idle(callable $callback, int $timeout = 1200, bool $auto_reconnect = false) {
-        $this->client->getConnection()->setConnectionTimeout($timeout);
-
-        $this->client->reconnect();
+    public function idle(callable $callback, int $timeout = 300, bool $auto_reconnect = false) {
+        $this->client->setTimeout($timeout);
         if (!in_array("IDLE", $this->client->getConnection()->getCapabilities())) {
             throw new NotSupportedCapabilityException("IMAP server does not support IDLE");
         }
         $this->client->openFolder($this->path, true);
         $connection = $this->client->getConnection();
+        $connection->idle();
 
         $sequence = ClientManager::get('options.sequence', IMAP::ST_MSGN);
-        $connection->idle();
 
         while (true) {
             try {
+                // This polymorphic call is fine - Protocol::idle() will throw an exception beforehand
                 $line = $connection->nextLine();
+
                 if (($pos = strpos($line, "EXISTS")) !== false) {
-                    $msgn = (int) substr($line, 2, $pos -2);
                     $connection->done();
+                    $msgn = (int) substr($line, 2, $pos -2);
 
                     $this->client->openFolder($this->path, true);
                     $message = $this->query()->getMessageByMsgn($msgn);
@@ -391,20 +391,26 @@ class Folder {
 
                     $event = $this->getEvent("message", "new");
                     $event::dispatch($message);
-
+                    $connection->idle();
+                } elseif (strpos($line, "OK") === false) {
+                    $connection->done();
                     $connection->idle();
                 }
             }catch (Exceptions\RuntimeException $e) {
+                if(strpos($e->getMessage(), "empty response") >= 0 && $connection->connected()) {
+                    $connection->done();
+                    $connection->idle();
+                    continue;
+                }
                 if(strpos($e->getMessage(), "connection closed") === false) {
                     throw $e;
                 }
-                if ($auto_reconnect === true) {
-                    $this->client->reconnect();
-                    $this->client->openFolder($this->path, true);
 
-                    $connection = $this->client->getConnection();
-                    $connection->idle();
-                }
+                $this->client->reconnect();
+                $this->client->openFolder($this->path, true);
+
+                $connection = $this->client->getConnection();
+                $connection->idle();
             }
         }
     }
@@ -428,7 +434,8 @@ class Folder {
      * @throws Exceptions\RuntimeException
      */
     public function examine(): array {
-        return $this->client->getConnection()->examineFolder($this->path);
+        $result = $this->client->getConnection()->examineFolder($this->path);
+        return is_array($result) ? $result : [];
     }
 
     /**